+++ lib/weird_phlex/component_pack/file.rb
@@ -3,61 +3,22 @@
 module WeirdPhlex
   class ComponentPack
     class File
-      attr_reader :component, :part, :file, :raw_file
+      attr_reader :component, :part, :file, :raw_file, :relative_path, :relative_path_without_part
 
-      def initialize(path, component_path:)
-        @path = path
-        @raw_file = path
-        @component_path = component_path
-        @relative_path = @path.to_s.delete_prefix("#{@component_path}/")
-        split = @relative_path.split('/')
+      def initialize(relative_path, component:)
+        @relative_path = relative_path
+        @relative_path_without_part = relative_path.gsub(%r(\A[^/]+/), '')
+        @component = component
+        @path = @component.component_path.join(relative_path)
+        @raw_file = @path
 
-        if ['version.rb', 'Rakefile'].include?(split.first)
-          @ignored = true
-          @shared_file = false
-          @component = nil
-          @part = nil
-          @file = split.first
-        elsif (matches = @relative_path.match(%r{\Ashared/(?<subdirectory>(?:[^/]+/)*)(?<component>_.*_)/(?<part>[^/]*)/(?<file>.*)}))
-          @shared_file = true
-          @component = nil
-          @part = matches[:part]
-          @file = matches[:file]
-        elsif (matches = @relative_path.match(%r{\A_global_/(?<part>[^/]*)/(?<file>.*)}))
-          @shared_file = true
-          @component = nil
-          @part = matches[:part]
-          @file = matches[:file]
-        elsif (matches = @relative_path.match(%r{\Acomponents/(?<subdirectory>(?:[^/]+/)*)(?<component>_.*_)/(?<part>[^/]*)/(?<file>.*)}))
-          @shared_file = false
-          @component = matches[:component]
-          @part = matches[:part]
-          @file = matches[:file]
-        else
-          raise "Regex error: could not parse file '#{@path}'"
-        end
+        matches = @relative_path.match(%r{\A(?<part>[^/]*)/(?<file>.*)})
+        @part = matches[:part]
+        @file = matches[:file]
       end
 
-      def ignored?
-        !!@ignored
-      end
-
-      def shared?
-        !@ignored && !!@shared_file
-      end
-
-      def component_file?
-        !@ignored && !@shared_file && @component.present?
-      end
-
-      def to_s
-        if @ignored
-          "IGNORED: #{@file}"
-        elsif @shared_file
-          "SHARED FILE: #{@relative_path}"
-        else
-          "FILE: #{@component} - #{@part} - #{@file}"
-        end
+      def inspect
+        "FILE #{@component.relative_path} - #{@relative_path}"
       end
     end
   end
