+++ lib/weird_phlex/planner/plan.rb
@@ -25,24 +25,6 @@ module WeirdPhlex
       end
 
       def create(source_file)
-        if source_file.shared?
-          add(
-            CopyAction.new(
-              source_file,
-              WeirdPhlex::Project::TargetFile.from_component_pack_shared_file(source_file),
-            ),
-          )
-        else
-          add(
-            CopyAction.new(
-              source_file,
-              WeirdPhlex::Project::TargetFile.from_component_pack_file(source_file),
-            ),
-          )
-        end
-      end
-
-      def create_2(source_file)
         add(
           CopyAction.new(
             source_file,
