+++ lib/weird_phlex/project/config.rb
@@ -8,18 +8,6 @@ module WeirdPhlex
         part_config('main').dig(component_type.to_s, part_name.to_s)
       end
 
-      def component_part_path(part_name)
-        part_config('main').dig('components', part_name.to_s)
-      end
-
-      def shared_part_path(shared_part_name)
-        part_config('main').dig('shared', shared_part_name.to_s)
-      end
-
-      def global_part_path(global_part_name)
-        part_config('main').dig('global', global_part_name.to_s)
-      end
-
       private
 
       def part_config(namespace)
